EB-005 draft purchase (#19)
* Draft purchase

* Draft purchase

* EB-006 | Happy path in tests

* EB-006 | Available Shipping for purchase

* EB-006 | Shipping Facade fix and tests added

* EB-005 fixes

* EB-005 fixes

+++ src/integration/kotlin/pl/exbook/exbook/ability/StockDomainAbility.kt
@@ -10,7 +10,7 @@ import pl.exbook.exbook.stock.adapter.rest.dto.StockDto
 import pl.exbook.exbook.utils.createHttpEntity
 
 class StockDomainAbility(private val restTemplate: TestRestTemplate) {
-    fun thereIsStock(startQuantity: Int = 100): ResponseEntity<StockDto> {
+    fun thereIsStock(startQuantity: Long = 100): ResponseEntity<StockDto> {
         val requestBody = CreateStockRequest(startQuantity)
         return restTemplate.postForEntity(
             "/api/stock",
