+++ src/main/kotlin/pl/exbook/exbook/stock/adapter/rest/dto/AddToStockRequest.kt
@@ -1,5 +1,5 @@
 package pl.exbook.exbook.stock.adapter.rest.dto
 
 data class AddToStockRequest(
-    val amount: Int
+    val amount: Long
 )
