+++ src/main/kotlin/pl/exbook/exbook/stock/domain/InsufficientStockException.kt
@@ -5,6 +5,6 @@ import pl.exbook.exbook.shared.ValidationException
 
 class InsufficientStockException(
     stockId: StockId,
-    requestedAmount: Int,
-    availableAmount: Int
+    requestedAmount: Long,
+    availableAmount: Long
 ) : ValidationException("Tried to reserve $requestedAmount for ${stockId.raw} stock amount but actual is $availableAmount")
